Move GeoInterface to an extension and refactored the rest to simplify code structure
