+++ src/time_parsing.jl
@@ -1,4 +1,4 @@
-module KMLTimeElementParsing
+module TimeParsing
 
 export parse_iso8601
 
@@ -348,114 +348,4 @@ matches_iso8601_format(s::AbstractString, format::Symbol) = begin
     getfield(@__MODULE__, validator)(s) === nothing
 end
 
-# Batch parsing for better performance
-"""
-    parse_iso8601_batch(strings::Vector{<:AbstractString}; warn::Bool=false) -> Vector
-
-Parse multiple ISO 8601 strings efficiently. Disables warnings by default for batch processing.
-"""
-function parse_iso8601_batch(strings::Vector{<:AbstractString}; warn::Bool=false)
-    [parse_iso8601(s, warn=warn) for s in strings]
-end
-
-# Test strings for ISO 8601 parser
-test_strings = [
-    # DateTime with timezone - Extended format
-    "2023-12-25T15:30:45Z",                    # UTC
-    "2023-12-25T15:30:45.123Z",                # UTC with milliseconds
-    "2023-12-25T15:30:45+01:00",               # Positive offset with colon
-    "2023-12-25T15:30:45-05:00",               # Negative offset with colon
-    "2023-12-25T15:30:45.500+02:30",           # With fractional seconds and offset
-    "2023-12-25 15:30:45Z",                    # Space separator instead of T
-    "2023-12-25T15:30:45America/New_York",     # Named timezone
-    "2023-12-25T15:30:45Europe/Paris",         # Named timezone
-    
-    # DateTime with timezone - Basic format
-    "20231225T153045Z",                        # UTC
-    "20231225T153045.123Z",                    # UTC with milliseconds
-    "20231225T153045+0100",                    # Positive offset without colon
-    "20231225T153045-0500",                    # Negative offset without colon
-    "20231225T153045.999+0230",                # With fractional seconds
-    
-    # DateTime without timezone - Extended format
-    "2023-12-25T15:30:45",                     # Standard format
-    "2023-12-25T15:30:45.123",                 # With milliseconds
-    "2023-12-25 15:30:45",                     # Space separator
-    "2023-12-25 15:30:45.999",                 # Space separator with milliseconds
-    
-    # DateTime without timezone - Basic format
-    "20231225T153045",                         # Standard format
-    "20231225T153045.123",                     # With milliseconds
-    
-    # Date only - Extended format
-    "2023-12-25",                              # Standard date
-    "2023-01-01",                              # New Year
-    "2023-02-28",                              # End of February (non-leap)
-    
-    # Date only - Basic format
-    "20231225",                                # Standard date
-    "20230101",                                # New Year
-    "20230228",                                # End of February
-    
-    # Week dates - Extended format
-    "2023-W52-1",                              # Week 52, Monday
-    "2023-W01-7",                              # Week 1, Sunday
-    "2023-W26-3",                              # Week 26, Wednesday
-    
-    # Week dates - Basic format
-    "2023W521",                                # Week 52, Monday
-    "2023W017",                                # Week 1, Sunday
-    "2023W263",                                # Week 26, Wednesday
-    
-    # Ordinal dates - Extended format
-    "2023-365",                                # Last day of year
-    "2023-001",                                # First day of year
-    "2023-180",                                # Middle of year
-    
-    # Ordinal dates - Basic format
-    "2023365",                                 # Last day of year
-    "2023001",                                 # First day of year
-    "2023180",                                 # Middle of year
-    
-    # Edge cases with valid formats
-    "2023-12-31T23:59:59Z",                   # End of year
-    "2023-01-01T00:00:00Z",                   # Start of year
-    "2023-12-31T23:59:60Z",                   # Leap second
-    "2023-12-25T15:30:45.123456789Z",         # Many decimal places
-    
-    # Invalid strings (should return as-is)
-    "not-a-date",
-    "2023-13-01",                              # Invalid month
-    "2023-12-32",                              # Invalid day
-    "2023-12-25T25:00:00",                    # Invalid hour
-    "202a-12-25",                              # Invalid year
-    "2023/12/25",                              # Wrong separator
-]
-
-# Test the parser
-"""
-    run_iso8601_tests(; strings=test_strings)
-
-Runs the ISO 8601 parser on a set of predefined test‐strings
-(or on a custom vector of strings passed via `strings`),
-prints each input together with its parsed result and type,
-and returns a Vector of (input, output) pairs.
-"""
-function run_iso8601_tests(; strings = [
-    "2023-12-25T15:30:45Z",
-    "2023-12-25T15:30:45.123Z",
-    "2023-12-25T15:30:45+01:00",
-    # … all the other test strings …
-    "not-a-date",
-    "2023-13-01",
-])
-    results = Vector{Tuple{String,Any}}(undef, length(strings))
-    for (i, s) in pairs(strings)
-        out = parse_iso8601(s)
-        # println("$s => $(typeof(out)): $out")
-        results[i] = (s, out)
-    end
-    return results
-end
-
-end # module KMLTimeElementParsing
+end # module TimeParsing
